Break out `PaperComponent` component from `UploadManager` to its own `DraggablePaper` file
